requirements.txt
altgraph==0.17.4
app==0.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
defusedxml==0.7.1
Flask==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jira==3.8.0
klembord==0.3.0
logging==0.4.9.6
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
python-dotenv==1.1.0
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
setuptools==78.1.0
soupsieve==2.7
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
